# api_spec.yaml
openapi: 3.1.0
info:
  title: MSR Event Hub API (Graph-aligned, merged + navigation properties)
  version: 0.3.0
  description: >
    Microsoft Graph–aligned resource shapes for MSR Event Hub entities and AI services.
    Conventions:
      - Resources use `id` and may include `@odata.type` and `@odata.etag`.
      - Collection responses use `{ "value": [...] }` and may include `@odata.nextLink`.
      - Error responses use `{ "error": { "code", "message", ... } }`.
      - Selected navigation properties are exposed as read-only navigation links and/or embedded objects.
servers:
  - url: https://eventhub.internal.microsoft.com
...
paths:
...
  /v1/events:
    get:
      tags: [Events]
      summary: List events
      parameters:
        - $ref: "#/components/parameters/Top"
        - $ref: "#/components/parameters/Skip"
        - $ref: "#/components/parameters/Filter"
        - $ref: "#/components/parameters/Select"
        - $ref: "#/components/parameters/OrderBy"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventCollectionResponse"
        default:
          $ref: "#/components/responses/GraphError"
    post:
      tags: [Events]
      summary: Create event (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventCreateRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        default:
          $ref: "#/components/responses/GraphError"
...
  /v1/events/{eventId}/projects:
    get:
      tags: [Projects]
      summary: List projects for an event
      parameters:
        - $ref: "#/components/parameters/EventId"
        - $ref: "#/components/parameters/Top"
        - $ref: "#/components/parameters/Skip"
        - $ref: "#/components/parameters/Filter"
        - $ref: "#/components/parameters/Select"
        - $ref: "#/components/parameters/OrderBy"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectCollectionResponse"
        default:
          $ref: "#/components/responses/GraphError"
    post:
      tags: [Projects]
      summary: Create project (admin or presenter draft)
      parameters:
        - $ref: "#/components/parameters/EventId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectCreateRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        default:
          $ref: "#/components/responses/GraphError"
...
  /v1/events/{eventId}/projects/{projectId}/knowledge:
    get:
      tags: [AI]
      summary: Get published project knowledge (human-approved)
      parameters:
        - $ref: "#/components/parameters/EventId"
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublishedProjectKnowledge"
        default:
          $ref: "#/components/responses/GraphError"
    put:
      tags: [AI]
      summary: Publish or update project knowledge (event admin / reviewer)
      parameters:
        - $ref: "#/components/parameters/EventId"
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublishedProjectKnowledgeUpsertRequest"
      responses:
        "200":
          description: Upserted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublishedProjectKnowledge"
        default:
          $ref: "#/components/responses/GraphError"
...
components:
  responses:
    GraphError:
      description: Error (Graph-style envelope)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/GraphErrorResponse"
  parameters:
    EventId:
      name: eventId
      in: path
      required: true
      schema: { $ref: "#/components/schemas/GraphId" }
...
    ProjectId:
      name: projectId
      in: path
      required: true
      schema: { $ref: "#/components/schemas/GraphId" }
    IfMatch:
      name: If-Match
      in: header
      required: false
      schema:
        type: string
      description: OData ETag for conditional update.
    Top:
      name: $top
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 999
    Skip:
      name: $skip
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        maximum: 1000000
    Filter:
      name: $filter
      in: query
      required: false
      schema:
        type: string
        maxLength: 4000
    Select:
      name: $select
      in: query
      required: false
      schema:
        type: string
        maxLength: 2000
      description: Comma-separated property names.
    OrderBy:
      name: $orderby
      in: query
      required: false
      schema:
        type: string
        maxLength: 2000
  schemas:
...
    GraphId:
      type: string
      minLength: 1
      maxLength: 256
      description: Unique identifier (Graph style).
    ODataType:
      type: string
      pattern: "^#.+\\..+$"
      description: OData type annotation.
    ODataEtag:
      type: string
      minLength: 1
      maxLength: 512
      description: OData ETag for concurrency.
    GraphDateTime:
      type: string
      format: date-time
      description: ISO 8601 date-time.
    GraphEntity:
      type: object
      additionalProperties: true
      required: [id]
      properties:
        "@odata.type":
          $ref: "#/components/schemas/ODataType"
        "@odata.etag":
          $ref: "#/components/schemas/ODataEtag"
        id:
          $ref: "#/components/schemas/GraphId"
    GraphCollectionResponse:
      type: object
      additionalProperties: true
      required: [value]
      properties:
        "@odata.context":
          type: string
          format: uri
        "@odata.nextLink":
          type: string
          format: uri
          description: Present when additional pages are available.
        value:
          type: array
          items: {}
    GraphErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: true
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            target:
              type: string
            details:
              type: array
              items:
                type: object
                additionalProperties: true
            innerError:
              type: object
              additionalProperties: true
    Url:
      type: string
      format: uri
      minLength: 8
    Email:
      type: string
      format: email
    Locale:
      type: string
      minLength: 2
      maxLength: 16
      pattern: "^[a-zA-Z]{2,3}(-[a-zA-Z0-9]{2,8})?$"

    # -------------------------
    # People & permissions
    # -------------------------
    Person:
      type: object
      additionalProperties: false
      required: [displayName]
      properties:
        id:
          $ref: "#/components/schemas/GraphId"
          description: Optional (e.g., Entra object id) if known.
        displayName:
          type: string
          minLength: 1
          maxLength: 200
        jobTitle:
          type: string
          maxLength: 200
        affiliation:
          type: string
          maxLength: 200
        mail:
          $ref: "#/components/schemas/Email"
        photoUrl:
          $ref: "#/components/schemas/Url"
        profileUrl:
          $ref: "#/components/schemas/Url"
    PermissionRole:
      type: string
      enum: [hubAdmin, eventAdmin, editor, presenter, viewer]
    PermissionBinding:
      type: object
      additionalProperties: false
      required: [role, principal]
      properties:
        role:
          $ref: "#/components/schemas/PermissionRole"
        principal:
          type: string
          minLength: 3
          maxLength: 256
          description: Opaque user/group identifier (e.g., Entra object id).

    # -------------------------
    # Assets & media
    # -------------------------
    AssetType:
      type: string
      enum: [paper, repo, deck, video, posterPdf, image, projectSite, dataset, other]
    Visibility:
      type: string
      enum: [internal, restricted]
      default: internal
    AssetLink:
      type: object
      additionalProperties: false
      required: [type, webUrl]
      properties:
        id:
          $ref: "#/components/schemas/GraphId"
        type:
          $ref: "#/components/schemas/AssetType"
        displayName:
          type: string
          maxLength: 200
        webUrl:
          $ref: "#/components/schemas/Url"
        visibility:
          $ref: "#/components/schemas/Visibility"
        notes:
          type: string
          maxLength: 2000
    Media:
      type: object
      additionalProperties: false
      properties:
        heroImageUrl:
          $ref: "#/components/schemas/Url"
        thumbnailUrl:
          $ref: "#/components/schemas/Url"
        posterPdfUrl:
          $ref: "#/components/schemas/Url"
        embeddedVideoUrl:
          $ref: "#/components/schemas/Url"

    # -------------------------
    # Hub configuration (homepage curation)
    # -------------------------
...
    EventStatus:
      type: string
      enum: [draft, published, archived]
    EventType:
      type: string
      enum: [researchShowcase, tab, workshop, lectureSeries, other]
    Track:
      type: object
      additionalProperties: false
      required: [id, displayName]
      properties:
        id:
          $ref: "#/components/schemas/GraphId"
        displayName:
          type: string
          maxLength: 200
        description:
          type: string
          maxLength: 2000
    Theme:
      type: object
      additionalProperties: false
      required: [id, displayName]
      properties:
        id:
          $ref: "#/components/schemas/GraphId"
        displayName:
          type: string
          maxLength: 200
        description:
          type: string
          maxLength: 2000
    Location:
      type: object
      additionalProperties: false
      properties:
        displayName:
          type: string
          maxLength: 200
        room:
          type: string
          maxLength: 80
        building:
          type: string
          maxLength: 80
        campus:
          type: string
          maxLength: 80
        address:
          type: string
          maxLength: 300
        floor:
          type: string
          maxLength: 40
        mapUrl:
          $ref: "#/components/schemas/Url"
    Event:
      allOf:
        - $ref: "#/components/schemas/GraphEntity"
        - type: object
          additionalProperties: false
          required: [displayName, status, type, timeZone, startDate, endDate]
          properties:
            "@odata.type":
              const: "#microsoft.graph.eventHubEvent"
            displayName:
              type: string
              minLength: 1
              maxLength: 200
            description:
              type: string
              maxLength: 50000
            status:
              $ref: "#/components/schemas/EventStatus"
            type:
              $ref: "#/components/schemas/EventType"
            timeZone:
              type: string
              minLength: 3
              maxLength: 64
              description: IANA timezone (e.g., Asia/Kolkata).
            locale:
              $ref: "#/components/schemas/Locale"
            startDate:
              type: string
              format: date
            endDate:
              type: string
              format: date
            venue:
              $ref: "#/components/schemas/Location"
            tracks:
              type: array
              items: { $ref: "#/components/schemas/Track" }
              default: []
            themes:
              type: array
              items: { $ref: "#/components/schemas/Theme" }
              default: []
            links:
              type: array
              items: { $ref: "#/components/schemas/AssetLink" }
              default: []
            permissions:
              type: array
              items: { $ref: "#/components/schemas/PermissionBinding" }
              default: []
            # ---- Navigation properties (Graph-style) ----
            sessions@odata.navigationLink:
              type: string
              format: uri
              readOnly: true
              description: URL to sessions collection for this event.
            projects@odata.navigationLink:
              type: string
              format: uri
              readOnly: true
              description: URL to projects collection for this event.
            bookmarks@odata.navigationLink:
              type: string
              format: uri
              readOnly: true
              description: URL to bookmarks collection for the signed-in user scoped to this event.
            createdDateTime:
              $ref: "#/components/schemas/GraphDateTime"
            lastModifiedDateTime:
              $ref: "#/components/schemas/GraphDateTime"
    EventCreateRequest:
      type: object
      additionalProperties: false
      required: [displayName, status, type, timeZone, startDate, endDate]
      properties:
        displayName:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 50000
        status:
          $ref: "#/components/schemas/EventStatus"
        type:
          $ref: "#/components/schemas/EventType"
        timeZone:
          type: string
          minLength: 3
          maxLength: 64
        locale:
          $ref: "#/components/schemas/Locale"
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        venue:
          $ref: "#/components/schemas/Location"
        tracks:
          type: array
          items: { $ref: "#/components/schemas/Track" }
          default: []
        themes:
          type: array
          items: { $ref: "#/components/schemas/Theme" }
          default: []
        links:
          type: array
          items: { $ref: "#/components/schemas/AssetLink" }
          default: []
        permissions:
          type: array
          items: { $ref: "#/components/schemas/PermissionBinding" }
          default: []
...
    EventCollectionResponse:
      allOf:
        - $ref: "#/components/schemas/GraphCollectionResponse"
        - type: object
          properties:
            value:
              type: array
              items: { $ref: "#/components/schemas/Event" }

    # -------------------------
    # Sessions
    # -------------------------
...
    ResearchMaturity:
      type: string
      enum: [exploratory, validated, pilotReady]
    ProjectLifecycleState:
      type: string
      enum: [draft, pendingReview, approved, published, archived]
    Project:
      allOf:
        - $ref: "#/components/schemas/GraphEntity"
        - type: object
          additionalProperties: false
          required: [eventId, title, abstract, team, maturity, lifecycleState]
          properties:
            "@odata.type":
              const: "#microsoft.graph.eventHubProject"
            eventId:
              $ref: "#/components/schemas/GraphId"
            humanReadableId:
              type: string
              minLength: 3
              maxLength: 40
              pattern: "^[A-Z0-9][A-Z0-9-]{2,39}$"
            title:
              type: string
              minLength: 1
              maxLength: 240
            abstract:
              type: string
              maxLength: 12000
            themeId:
              $ref: "#/components/schemas/GraphId"
            trackId:
              $ref: "#/components/schemas/GraphId"
            location:
              $ref: "#/components/schemas/Location"
            team:
              type: array
              minItems: 1
              maxItems: 50
              items: { $ref: "#/components/schemas/Person" }
            primaryContact:
              $ref: "#/components/schemas/Person"
            maturity:
              $ref: "#/components/schemas/ResearchMaturity"
            lifecycleState:
              $ref: "#/components/schemas/ProjectLifecycleState"
            media:
              $ref: "#/components/schemas/Media"
            assets:
              type: array
              items: { $ref: "#/components/schemas/AssetLink" }
              default: []
            visibility:
              $ref: "#/components/schemas/Visibility"
            # ---- Navigation properties (Graph-style) ----
            poster@odata.navigationLink:
              type: string
              format: uri
              readOnly: true
              description: URL to poster resource for this project.
            knowledge@odata.navigationLink:
              type: string
              format: uri
              readOnly: true
              description: URL to published knowledge resource for this project.
            # Optional embedded navigation objects (read-only)
            poster:
              $ref: "#/components/schemas/Poster"
              readOnly: true
            publishedKnowledge:
              $ref: "#/components/schemas/PublishedProjectKnowledge"
              readOnly: true
            createdDateTime:
              $ref: "#/components/schemas/GraphDateTime"
            lastModifiedDateTime:
              $ref: "#/components/schemas/GraphDateTime"
    ProjectCreateRequest:
      type: object
      additionalProperties: false
      required: [eventId, title, abstract, team, maturity, lifecycleState]
      properties:
        eventId:
          $ref: "#/components/schemas/GraphId"
        humanReadableId:
          type: string
          maxLength: 40
        title:
          type: string
          minLength: 1
          maxLength: 240
        abstract:
          type: string
          maxLength: 12000
        themeId:
          $ref: "#/components/schemas/GraphId"
        trackId:
          $ref: "#/components/schemas/GraphId"
        location:
          $ref: "#/components/schemas/Location"
        team:
          type: array
          minItems: 1
          maxItems: 50
          items: { $ref: "#/components/schemas/Person" }
        primaryContact:
          $ref: "#/components/schemas/Person"
        maturity:
          $ref: "#/components/schemas/ResearchMaturity"
        lifecycleState:
          $ref: "#/components/schemas/ProjectLifecycleState"
          description: For presenter self-service, starts as draft/pendingReview.
        media:
          $ref: "#/components/schemas/Media"
        assets:
          type: array
          items: { $ref: "#/components/schemas/AssetLink" }
          default: []
        visibility:
          $ref: "#/components/schemas/Visibility"
...
    ProjectCollectionResponse:
      allOf:
        - $ref: "#/components/schemas/GraphCollectionResponse"
        - type: object
          properties:
            value:
              type: array
              items: { $ref: "#/components/schemas/Project" }

    # -------------------------
    # Bookmark (QR + personal lists)
    # -------------------------
...
    Poster:
      allOf:
        - $ref: "#/components/schemas/GraphEntity"
        - type: object
          additionalProperties: false
          required: [eventId, projectId, posterPdfUrl]
          properties:
            "@odata.type":
              const: "#microsoft.graph.eventHubPoster"
            eventId:
              $ref: "#/components/schemas/GraphId"
            projectId:
              $ref: "#/components/schemas/GraphId"
            posterPdfUrl:
              $ref: "#/components/schemas/Url"
              description: Canonical poster PDF.
            derived:
              type: object
              additionalProperties: false
              properties:
                heroImageUrl:
                  $ref: "#/components/schemas/Url"
                thumbnailUrl:
                  $ref: "#/components/schemas/Url"
            embeddedVideoUrl:
              $ref: "#/components/schemas/Url"
            createdDateTime:
              $ref: "#/components/schemas/GraphDateTime"
            lastModifiedDateTime:
              $ref: "#/components/schemas/GraphDateTime"
...
    KnowledgeSourceRef:
      type: object
      additionalProperties: false
      required: [snapshotId, sourceUri]
      properties:
        snapshotId:
          type: string
          maxLength: 200
        sourceUri:
          $ref: "#/components/schemas/Url"
        capturedDateTime:
          $ref: "#/components/schemas/GraphDateTime"
        contentHash:
          type: string
          maxLength: 200
        notes:
          type: string
          maxLength: 1000
    PublishedKnowledgeSection:
      type: object
      additionalProperties: false
      required: [text, citations]
      properties:
        text:
          type: string
          maxLength: 8000
        citations:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/PKACitation"
    PublishedProjectKnowledge:
      allOf:
        - $ref: "#/components/schemas/GraphEntity"
        - type: object
          additionalProperties: false
          required:
            - eventId
            - projectId
            - schemaVersion
            - approvedBy
            - approvedDateTime
            - sources
            - heilmeier
            - faq
          properties:
            "@odata.type":
              const: "#microsoft.graph.eventHubPublishedProjectKnowledge"
            eventId:
              $ref: "#/components/schemas/GraphId"
            projectId:
              $ref: "#/components/schemas/GraphId"
            schemaVersion:
              type: string
              const: "published-1.0"
            approvedBy:
              type: string
              minLength: 3
              maxLength: 256
            approvedDateTime:
              $ref: "#/components/schemas/GraphDateTime"
            sourceDraftArtifactIds:
              type: array
              items: { $ref: "#/components/schemas/GraphId" }
              default: []
            sources:
              type: array
              minItems: 1
              items: { $ref: "#/components/schemas/KnowledgeSourceRef" }
            heilmeier:
              type: object
              additionalProperties: false
              required:
                - whatAreYouTryingToDo
                - howIsItDoneToday
                - whatIsNew
                - whoCares
                - risksAndPayoffs
                - metricsOfSuccess
              properties:
                whatAreYouTryingToDo:
                  $ref: "#/components/schemas/PublishedKnowledgeSection"
                howIsItDoneToday:
                  $ref: "#/components/schemas/PublishedKnowledgeSection"
                whatIsNew:
                  $ref: "#/components/schemas/PublishedKnowledgeSection"
                whoCares:
                  $ref: "#/components/schemas/PublishedKnowledgeSection"
                risksAndPayoffs:
                  $ref: "#/components/schemas/PublishedKnowledgeSection"
                metricsOfSuccess:
                  $ref: "#/components/schemas/PublishedKnowledgeSection"
            faq:
              type: array
              default: []
              items:
                type: object
                additionalProperties: false
                required: [question, answer]
                properties:
                  question:
                    type: string
                    maxLength: 300
                  answer:
                    $ref: "#/components/schemas/PublishedKnowledgeSection"
            scopeNotes:
              type: string
              maxLength: 4000
            createdDateTime:
              $ref: "#/components/schemas/GraphDateTime"
            lastModifiedDateTime:
              $ref: "#/components/schemas/GraphDateTime"
    PublishedProjectKnowledgeUpsertRequest:
      type: object
      additionalProperties: false
      required: [approvedBy, approvedDateTime, sources, heilmeier, faq]
      properties:
        approvedBy:
          type: string
          minLength: 3
          maxLength: 256
        approvedDateTime:
          $ref: "#/components/schemas/GraphDateTime"
        sourceDraftArtifactIds:
          type: array
          items: { $ref: "#/components/schemas/GraphId" }
          default: []
        sources:
          type: array
          minItems: 1
          items: { $ref: "#/components/schemas/KnowledgeSourceRef" }
        heilmeier:
          $ref: "#/components/schemas/PublishedProjectKnowledge/properties/heilmeier"
        faq:
          $ref: "#/components/schemas/PublishedProjectKnowledge/properties/faq"
        scopeNotes:
          type: string
          maxLength: 4000